search - move appointment card into its own file
